Make template implementation in separate header, making it testable and exportable
