+++ exes/ethercat/inc/public/tfc/ec/devices/beckhoff/EL2xxx.hpp
@@ -2,6 +2,7 @@
 
 #include <bitset>
 #include <cstdint>
+#include <cstddef>
 #include <memory>
 #include <span>
 #include <vector>
