+++ exes/ethercat/tests/devices/beckhoff.cpp
@@ -1,8 +1,8 @@
 #include <boost/asio.hpp>
 #include <boost/ut.hpp>
 
-#include <tfc/ec/devices/beckhoff/EL1xxx.hpp>
-#include <tfc/ec/devices/beckhoff/EL2xxx.hpp>
+#include <tfc/ec/devices/beckhoff/EL1xxx_impl.hpp>
+#include <tfc/ec/devices/beckhoff/EL2xxx_impl.hpp>
 #include <tfc/ipc.hpp>
 #include <tfc/ipc/details/dbus_server_iface_mock.hpp>
 #include <tfc/progbase.hpp>
